+++ src/main/client/query.cc
@@ -109,83 +109,70 @@ void ParseWhereClause(as_query* query, Local<Object> filter, LogInfo* log)
     // Parse the filters and set the filters to query object
     if ( filter->IsArray() ) {
         Local<Array> filters = Local<Array>::Cast(filter);
-        int size             = filters->Length();
-        as_v8_debug(log, "Number of filters %d", size);
+        int size = filters->Length();
+        as_v8_debug(log, "Number of filters: %d", size);
         as_query_where_init(query, (uint16_t)size);
         for (int i=0; i < size; i++) {
             Local<Object> filter = filters->Get(i)->ToObject();
-            Local<Value> bin     = filter->Get(Nan::New("bin").ToLocalChecked());
-            char * bin_val       = NULL;
-            if( !bin->IsString() ) {
+            Local<Value> bin = filter->Get(Nan::New("bin").ToLocalChecked());
+            if (!bin->IsString()) {
                 as_v8_error(log, "Bin value must be string");
                 Nan::ThrowError("Bin value is not a string");
             }
-            int predicate        = filter->Get(Nan::New("predicate").ToLocalChecked())->ToObject()->IntegerValue();
-            as_v8_debug(log, "Bin name in the filter %s \n", *String::Utf8Value(bin));
-            switch(predicate)
-            {
-                case AS_PREDICATE_RANGE:
-                    {
-                        as_index_datatype type = (as_index_datatype)filter->Get(Nan::New("type").ToLocalChecked())->ToObject()->IntegerValue();
-                        if( type == AS_INDEX_NUMERIC) {
-                            Local<Value> v8min = filter->Get(Nan::New("min").ToLocalChecked());
-                            Local<Value> v8max = filter->Get(Nan::New("max").ToLocalChecked());
-                            int64_t min = 0, max = 0;
-                            if( v8min->IsNumber()) {
-                                min = v8min->NumberValue();
-                            }
-                            else {
-                                as_v8_error(log, "The range value passed must be an integer");
-                                Nan::ThrowError("The range value passed is not an integer");
-                            }
-                            if( v8max->IsNumber()){
-                                max = v8max->NumberValue();
-                            }
-                            else {
-                                as_v8_error(log, "The range value passed must be an integer");
-                                Nan::ThrowError("The range value passed is not an integer");
-                            }
-                            as_query_where( query, *String::Utf8Value(bin), as_integer_range(min, max));
-                            as_v8_debug(log, "Integer range predicate from %d to %d", min, max);
-                            break;
+            const char* bin_name = strdup(*String::Utf8Value(bin));
+            as_predicate_type predicate = (as_predicate_type) filter->Get(Nan::New("predicate").ToLocalChecked())->ToObject()->IntegerValue();
+            as_index_type type = (as_index_type) filter->Get(Nan::New("type").ToLocalChecked())->ToObject()->IntegerValue();
+            as_index_datatype datatype = (as_index_datatype) filter->Get(Nan::New("datatype").ToLocalChecked())->ToObject()->IntegerValue();
+            as_v8_debug(log, "Building filter on predicate type %d, index type %d, data type %d, bin name '%s'", predicate, type, datatype, bin_name)
+            switch(predicate) {
+                case AS_PREDICATE_RANGE: {
+                    if (datatype == AS_INDEX_NUMERIC) {
+                        Local<Value> v8min = filter->Get(Nan::New("min").ToLocalChecked());
+                        Local<Value> v8max = filter->Get(Nan::New("max").ToLocalChecked());
+                        if (v8min->IsNumber() && v8max->IsNumber()) {
+                            const int64_t min = v8min->NumberValue();
+                            const int64_t max = v8max->NumberValue();
+                            as_query_where(query, bin_name, predicate, type, datatype, min, max);
+                            as_v8_debug(log, "Integer range predicate from %llu to %llu", min, max);
+                        } else {
+                            as_v8_error(log, "The min/max of the range value passed must both be integers.");
+                            Nan::ThrowError("The min/max of the range value passed must both be integers.");
                         }
-                        else if(type == AS_INDEX_GEO2DSPHERE) {
-                            Local<Value> val = filter->Get(Nan::New("val").ToLocalChecked());
-                            if( !val->IsString()) {
-                                as_v8_error(log, "The region value passed must be a GeoJSON string");
-                                Nan::ThrowError("The region value passed is not a GeoJSON string");
-                            }
-                            bin_val = strdup(*String::Utf8Value(val));
-                            as_query_where(query, *String::Utf8Value(bin), as_geo_within(bin_val));
-                            as_v8_debug(log, " Geo range predicate %s", bin_val);
-                            break;
+                    } else if (datatype == AS_INDEX_GEO2DSPHERE) {
+                        Local<Value> value = filter->Get(Nan::New("val").ToLocalChecked());
+                        if (!value->IsString()) {
+                            as_v8_error(log, "The region value passed must be a GeoJSON string");
+                            Nan::ThrowError("The region value passed is not a GeoJSON string");
                         }
+                        const char* bin_val = strdup(*String::Utf8Value(value));
+                        as_query_where(query, bin_name, predicate, type, datatype, bin_val);
+                        as_v8_debug(log, "Geo range predicate %s", bin_val);
                     }
-                case AS_PREDICATE_EQUAL:
-                    {
-                        as_index_datatype type = (as_index_datatype)filter->Get(Nan::New("type").ToLocalChecked())->ToObject()->IntegerValue();
-                        if( type == AS_INDEX_NUMERIC) {
-                            if( !filter->Get(Nan::New("val").ToLocalChecked())->IsNumber()) {
-                                as_v8_error(log, "querying an integer index with equal predicate - value must be an integer");
-                                Nan::ThrowError("Querying an integer index with equal predicate - value is not an integer");
-                            }
-                            int64_t val = filter->Get(Nan::New("val").ToLocalChecked())->ToObject()->NumberValue();
-                            as_query_where( query, *String::Utf8Value(bin), as_integer_equals(val));
-                            as_v8_debug(log," Integer equality predicate %llu", val);
-                            break;
+                    break;
+                }
+                case AS_PREDICATE_EQUAL: {
+                    if (datatype == AS_INDEX_NUMERIC) {
+                        Local<Value> value = filter->Get(Nan::New("val").ToLocalChecked());
+                        if (value->IsNumber()) {
+                            const int64_t val = value->NumberValue();
+                            as_query_where(query, bin_name, predicate, type, datatype, val);
+                            as_v8_debug(log, "Integer equality predicate %d", val);
+                        } else {
+                            as_v8_error(log, "querying a numeric index with equal predicate - value must be a number");
+                            Nan::ThrowError("Querying an numeric index with equal predicate - value is not a number");
                         }
-                        else if(type == AS_INDEX_STRING) {
-                            Local<Value> val = filter->Get(Nan::New("val").ToLocalChecked());
-                            if( !val->IsString()) {
-                                as_v8_error(log," querying a string index with equal predicate - value must be a string");
-                                Nan::ThrowError("Querying a string index with equal predicate - value is not a string");
-                            }
-                            bin_val   = strdup(*String::Utf8Value(val));
-                            as_query_where( query, *String::Utf8Value(bin),as_string_equals(bin_val));
-                            as_v8_debug(log, " String equality predicate %s", bin_val);
-                            break;
+                    } else if (datatype == AS_INDEX_STRING) {
+                        Local<Value> value = filter->Get(Nan::New("val").ToLocalChecked());
+                        if (!value->IsString()) {
+                            as_v8_error(log, "querying a string index with equal predicate - value must be a string");
+                            Nan::ThrowError("Querying a string index with equal predicate - value is not a string");
                         }
+                        const char* bin_val = strdup(*String::Utf8Value(value));
+                        as_query_where(query, bin_name, predicate, type, datatype, bin_val);
+                        as_v8_debug(log, "String equality predicate %s", bin_val);
                     }
+                    break;
+                }
             }
         }
     }