+++ test/generators/value.js
@@ -17,10 +17,32 @@
 // *****************************************************************************
 // HELPERS
 // ****************************************************************************
+//
+const Aerospike = require('../../lib/aerospike')
+const Double = Aerospike.Double
+const GeoJSON = Aerospike.GeoJSON
 
-function random (min, max) {
-  var rand = Math.floor(Math.random() * 0x100000000) % max
-  return rand < min ? min : rand
+// Returns a random integer between min (included) and max (excluded)
+function randomInt (min, max) {
+  return Math.floor(Math.random() * (max - min)) + min
+}
+
+// Returns a random number between min (included) and max (excluded)
+function randomDouble (min, max) {
+  return Math.random() * (max - min) + min
+}
+
+// Returns a random point in a circle of radius r (in meters) around the
+// geographical coordiates lat, lng
+// Source: http://gis.stackexchange.com/a/25883/10736
+function randomPoint (lat, lng, r) {
+  r = r / 111300 // radius in degrees
+  var w = r * Math.sqrt(Math.random())
+  var t = 2 * Math.PI * Math.random()
+  var x = w * Math.cos(t)
+  var y = w * Math.sin(t)
+  x = x / Math.cos(lng) // adjust for shrinking of east-west distances
+  return [lat + x, lng + y]
 }
 
 function merge (o1, o2) {
@@ -54,10 +76,10 @@ function string (options) {
   var seq = 0
   return function () {
     if (opt.random === true) {
-      var len = random(opt.length.min, opt.length.max)
+      var len = randomInt(opt.length.min, opt.length.max)
       var arr = new Array(len)
       for (var i = 0; i < len; i++) {
-        arr[i] = opt.charset[random(0, opt.charset.length)]
+        arr[i] = opt.charset[randomInt(0, opt.charset.length)]
       }
       return opt.prefix + arr.join('') + opt.suffix
     } else {
@@ -79,10 +101,10 @@ string.defaults = {
 function bytes (options) {
   var opt = merge(bytes.defaults, options)
   return function () {
-    var len = random(opt.length.min, opt.length.max)
+    var len = randomInt(opt.length.min, opt.length.max)
     var buf = new Buffer(len)
     for (var i = 0; i < len; i++) {
-      buf[i] = random(opt.byte.min, opt.byte.max)
+      buf[i] = randomInt(opt.byte.min, opt.byte.max)
     }
     return buf
   }
@@ -103,23 +125,60 @@ function integer (options) {
   var opt = merge(integer.defaults, options)
   var seq = opt.min
   return function () {
-    return opt.random === true ? random(opt.min, opt.max) : seq++
+    return opt.random === true ? randomInt(opt.min, opt.max) : seq++
   }
 }
+
 integer.defaults = {
   random: true,
   min: 0,
   max: 0xffffff
 }
 
-function array () {
+function double (options) {
+  var opt = merge(double.defaults, options)
+  var seq = opt.min
+  var step = opt.step
+  var r = Math.pow(10, step.toString().length - step.toString().indexOf('.') - 1)
   return function () {
-    var itype = integer()
-    var stype = string()
-    var btype = bytes()
-    var arr = [itype(), stype(), btype()]
-    return arr
+    if (opt.random) {
+      return new Double(randomDouble(opt.min, opt.max))
+    } else {
+      seq = Math.round(r * (seq + step)) / r
+      return new Double(seq)
+    }
   }
+}
+
+double.defaults = {
+  random: true,
+  min: 0,
+  max: 0xffffff,
+  step: 0.1
+}
+
+function geojsonPoint (options) {
+  var opt = merge(geojsonPoint.defaults, options)
+  return function () {
+    var coords = (opt.random === true) ? randomPoint(opt.lat, opt.lng, opt.r) : [opt.lat, opt.lng]
+    return new GeoJSON({type: 'Point', coordinates: coords.reverse()})
+  }
+}
+geojsonPoint.defaults = {
+  random: true,
+  lat: 37.4214209,
+  lng: -122.1008744,
+  r: 1000
+}
+
+function array (options) {
+  var opt = merge(array.defaults, options)
+  return function () {
+    return opt.values.map(function (gen) { return gen() })
+  }
+}
+array.defaults = {
+  values: [integer(), string(), bytes()]
 }
 
 function map () {
@@ -155,6 +214,8 @@ module.exports = {
   constant: constant,
   integer: integer,
   string: string,
+  double: double,
+  geojsonPoint: geojsonPoint,
   array: array,
   map: map,
   array_of_array: array_of_array,